Began working on a login/password system, but this is far from secure atm.
For example, you can run the following in the console and log right in:

C.user_preferences.addons['maze_gen'].preferences['access'] = True

...because the 'access' property is accessed using the key/dict notation instead of dot notation (...preferences.access = True). I'll need to make this non-editable with a set function, then write another function to change its value by setting some other internal encryption to some value that allows an update with the set function.

Note that the setter is for ['access'] notation and update is for .access notation. Getters are then simply used for reading (I believe the draw methods will use this as well as just accessing it from the API with print(...prefs.access) or print(...prefs['access'])).